+++ 0.5/Control/Monad/Trans/Accum.hs
@@ -3,17 +3,19 @@
 #ifndef HASKELL98
 {-# LANGUAGE DeriveDataTypeable #-}
 {-# LANGUAGE StandaloneDeriving #-}
-# if __GLASGOW_HASKELL__ >= 704
-{-# LANGUAGE Safe #-}
-# elif __GLASGOW_HASKELL__ >= 702
-{-# LANGUAGE Trustworthy #-}
-# endif
+{-# LANGUAGE TypeFamilies #-}
 # if __GLASGOW_HASKELL__ >= 706
 {-# LANGUAGE PolyKinds #-}
 # endif
 # if __GLASGOW_HASKELL__ >= 708
 {-# LANGUAGE AutoDeriveTypeable #-}
 {-# LANGUAGE DataKinds #-}
+# endif
+# if __GLASGOW_HASKELL__ >= 710
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE Safe #-}
+# elif __GLASGOW_HASKELL__ >= 702
+{-# LANGUAGE Trustworthy #-}
 # endif
 #endif
 -----------------------------------------------------------------------------
@@ -84,6 +86,14 @@ import Data.Monoid
 import Data.Typeable
 #endif
 
+#ifndef HASKELL98
+# ifdef GENERIC_DERIVING
+import Generics.Deriving.Base
+# elif __GLASGOW_HASKELL__ >= 702
+import GHC.Generics
+# endif
+#endif
+
 -- ---------------------------------------------------------------------------
 -- | An accumulation monad parameterized by the type @w@ of output to accumulate.
 --
@@ -143,6 +153,30 @@ mapAccum f = mapAccumT (Identity . f . runIdentity)
 --
 --  * a writer monad transformer with the extra ability to read all previous output.
 newtype AccumT w m a = AccumT (w -> m (a, w))
+
+#ifndef HASKELL98
+# if __GLASGOW_HASKELL__ >= 710
+deriving instance Generic (AccumT w m a)
+# elif __GLASGOW_HASKELL__ >= 702 || defined(GENERIC_DERIVING)
+instance Generic (AccumT w m a) where
+  type Rep (AccumT w m a) = D1 D1'AccumT (C1 C1_0'AccumT (S1 NoSelector (Rec0 (w -> m (a, w)))))
+  from (AccumT x) = M1 (M1 (M1 (K1 x)))
+  to (M1 (M1 (M1 (K1 x)))) = AccumT x
+
+instance Datatype D1'AccumT where
+  datatypeName _ = "AccumT"
+  moduleName _ = "Control.Monad.Trans.Accum"
+#  if MIN_VERSION_base(4,7,0)
+  isNewtype _ = True
+#  endif
+
+instance Constructor C1_0'AccumT where
+  conName _ = "AccumT"
+
+data D1'AccumT
+data C1_0'AccumT
+# endif
+#endif
 
 -- | Unwrap an accumulation computation.
 runAccumT :: AccumT w m a -> w -> m (a, w)
@@ -205,8 +239,10 @@ instance (Monoid w, Functor m, Monad m) => Monad (AccumT w m) where
         ~(b, w'') <- runAccumT (k a) (w `mappend` w')
         return (b, w' `mappend` w'')
     {-# INLINE (>>=) #-}
+#if !(MIN_VERSION_base(4,13,0))
     fail msg = AccumT $ const (fail msg)
     {-# INLINE fail #-}
+#endif
 
 instance (Monoid w, Functor m, Fail.MonadFail m) => Fail.MonadFail (AccumT w m) where
     fail msg = AccumT $ const (Fail.fail msg)